requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Slack
slack-sdk==3.26.2
aiohttp==3.9.1

# HTTP client for Grok
httpx==0.26.0

# Redis
redis==5.0.1

# Security (for Slack signature verification)
cryptography==42.0.0
